refactor: make Targets traceable
Summary:
# Changelog: [Internal]

The current design for Tracing is flawed. Right now the logic is mostly scattered around CDP Agents, lifetime of which is tied to CDP session.

This diff introduces a new approach: Targets will expose an API to start / stop tracing.

This allows us to record Tracing Profiles even if there is no active CDP session.

Differential Revision: D78517818