Use nearest component name as default debugID for AnimatedValue
Summary:
## Changelog:

[Internal] [Added] - Use nearest component name as default debugID for AnimatedValue

Motivation: 

Debugging animation was difficult partially because narrowing down to the culprit AnimatedNode was tricky. I once added debugID to animatedNode config, but it can still take a lot of manual work to add debugID to all possible AnimatedNodes

So here i'm trying to use component name where the AnimatedValue is created in as default debugID - the approach is by parsing js error stack at constructor of AnimatedValue. Some example stack traces: P1876475287

Currently I'm only adding this to AnimatedValue and interpolation because these seem to be most used in product

Differential Revision: D78584306