Isolate NativeDevSettings load
Summary:
While adding an opt native/dev js mode to fantom, LogBox was trying to load NativeDevSettings when it didn't exist because of mode mismatch.

This is the only location it's happening, so this just moved the NativeDevSettings load into the one function that needs it, probably not the best solution because if this function gets called in this mode we will crash still, but it works for now.

Changelog: [Internal]

Differential Revision: D78952284