fix crash where C++ Animated generates update for deleted view (#52804)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/52804

changelog: [internal]

On Android, we can run into this problem where a deleted view is being animated. This occurs specifically when UI tick occurs after a view is deleted but before an animation for the view is disconnected.

How come iOS doesn't crash?
It has a protection against this: https://fburl.com/code/pdoexj34

Reviewed By: zeyap

Differential Revision: D78815390