Factor out accessibility props copying for View component (#52899)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/52899

## Changelog
[Internal] -

This factors out the code for populating Accessibility props for a View component (which is now done in an optimized way, avoiding populating the default values altogether).

This is done in a generic way, with the goal of reusing the same code for other component types (such as Text and Image).

Differential Revision: D79163459